+++ FWFramework/Classes/Compatible/Kernel/Navigation.swift
@@ -41,14 +41,14 @@ extension Wrapper where Base: UIWindow {
     }
 
     /// 使用最顶部的视图控制器打开控制器，自动判断push|present
-    public func open(_ viewController: UIViewController, animated: Bool = true) {
-        base.__fw_open(viewController, animated: animated)
+    public func open(_ viewController: UIViewController, animated: Bool = true, completion: (() -> Void)? = nil) {
+        base.__fw_open(viewController, animated: animated, completion: completion)
     }
 
     /// 关闭最顶部的视图控制器，自动判断pop|dismiss，返回是否成功
     @discardableResult
-    public func close(animated: Bool = true) -> Bool {
-        return base.__fw_closeViewController(animated: true)
+    public func close(animated: Bool = true, completion: (() -> Void)? = nil) -> Bool {
+        return base.__fw_closeViewController(animated: true, completion: completion)
     }
     
     // MARK: - Static
@@ -90,14 +90,14 @@ extension Wrapper where Base: UIWindow {
     }
 
     /// 使用最顶部的视图控制器打开控制器，自动判断push|present
-    public static func open(_ viewController: UIViewController, animated: Bool = true) {
-        Base.__fw_open(viewController, animated: animated)
+    public static func open(_ viewController: UIViewController, animated: Bool = true, completion: (() -> Void)? = nil) {
+        Base.__fw_open(viewController, animated: animated, completion: completion)
     }
 
     /// 关闭最顶部的视图控制器，自动判断pop|dismiss，返回是否成功
     @discardableResult
-    public static func close(animated: Bool = true) -> Bool {
-        return Base.__fw_closeViewController(animated: true)
+    public static func close(animated: Bool = true, completion: (() -> Void)? = nil) -> Bool {
+        return Base.__fw_closeViewController(animated: true, completion: completion)
     }
     
 }
@@ -105,15 +105,22 @@ extension Wrapper where Base: UIWindow {
 // MARK: - UIViewController+Navigation
 extension Wrapper where Base: UIViewController {
     
+    // MARK: - Navigation
+    /// 自定义open|close导航样式，默认automatic自动判断
+    public var navigationOptions: NavigationOptions {
+        get { return base.__fw_navigationOptions }
+        set { base.__fw_navigationOptions = newValue }
+    }
+    
     /// 打开控制器。1.如果打开导航栏，则调用present；2.否则如果导航栏存在，则调用push；3.否则调用present
-    public func open(_ viewController: UIViewController, animated: Bool = true) {
-        base.__fw_open(viewController, animated: animated)
+    public func open(_ viewController: UIViewController, animated: Bool = true, completion: (() -> Void)? = nil) {
+        base.__fw_open(viewController, animated: animated, completion: completion)
     }
 
     /// 关闭控制器，返回是否成功。1.如果导航栏不存在，则调用dismiss；2.否则如果已是导航栏底部，则调用dismiss；3.否则调用pop
     @discardableResult
-    public func close(animated: Bool = true) -> Bool {
-        return base.__fw_close(animated: true)
+    public func close(animated: Bool = true, completion: (() -> Void)? = nil) -> Bool {
+        return base.__fw_close(animated: true, completion: completion)
     }
     
     // MARK: - Workflow
@@ -128,6 +135,35 @@ extension Wrapper where Base: UIViewController {
 // MARK: - UINavigationController+Navigation
 extension Wrapper where Base: UINavigationController {
     
+    // MARK: - Navigation
+    /// push新界面，完成时回调
+    public func pushViewController(_ viewController: UIViewController, animated: Bool, completion: (() -> Void)? = nil) {
+        base.__fw_pushViewController(viewController, animated: animated, completion: completion)
+    }
+
+    /// pop当前界面，完成时回调
+    @discardableResult
+    public func popViewController(animated: Bool, completion: (() -> Void)? = nil) -> UIViewController? {
+        return base.__fw_popViewController(animated: animated, completion: completion)
+    }
+
+    /// pop到指定界面，完成时回调
+    @discardableResult
+    public func popToViewController(_ viewController: UIViewController, animated: Bool, completion: (() -> Void)? = nil) -> [UIViewController]? {
+        return base.__fw_pop(to: viewController, animated: animated, completion: completion)
+    }
+
+    /// pop到根界面，完成时回调
+    @discardableResult
+    public func popToRootViewController(animated: Bool, completion: (() -> Void)? = nil) -> [UIViewController]? {
+        return base.__fw_popToRootViewController(animated: animated, completion: completion)
+    }
+
+    /// 设置界面数组，完成时回调
+    public func setViewControllers(_ viewControllers: [UIViewController], animated: Bool, completion: (() -> Void)? = nil) {
+        base.__fw_setViewControllers(viewControllers, animated: animated, completion: completion)
+    }
+    
     // MARK: - Workflow
     /// 当前最外层工作流名称，即topViewController的工作流名称
     public var topWorkflowName: String? {
@@ -137,36 +173,36 @@ extension Wrapper where Base: UINavigationController {
     /// push控制器，并清理最外层工作流（不属于工作流则不清理）
     ///
     /// 示例：1、（2、3）、4、（5、6）、（7、8），操作后为1、（2、3）、4、（5、6）、9
-    public func push(_ viewController: UIViewController, popTopWorkflowAnimated: Bool) {
-        base.__fw_pushViewController(viewController, popTopWorkflowAnimated: popTopWorkflowAnimated)
+    public func push(_ viewController: UIViewController, popTopWorkflowAnimated: Bool, completion: (() -> Void)? = nil) {
+        base.__fw_pushViewController(viewController, popTopWorkflowAnimated: popTopWorkflowAnimated, completion: completion)
     }
 
     /// push控制器，并清理非根控制器（只保留根控制器）
     ///
     /// 示例：1、（2、3）、4、（5、6）、（7、8），操作后为1、9
-    public func push(_ viewController: UIViewController, popToRootWorkflowAnimated: Bool) {
-        base.__fw_pushViewController(viewController, popToRootWorkflowAnimated: popToRootWorkflowAnimated)
+    public func push(_ viewController: UIViewController, popToRootWorkflowAnimated: Bool, completion: (() -> Void)? = nil) {
+        base.__fw_pushViewController(viewController, popToRootWorkflowAnimated: popToRootWorkflowAnimated, completion: completion)
     }
 
     /// push控制器，并从外到内清理指定工作流，直到遇到不属于指定工作流的控制器停止
     ///
     /// 示例：1、（2、3）、4、（5、6）、（7、8），操作后为1、（2、3）、4、9
-    public func push(_ viewController: UIViewController, popWorkflows: [String]?, animated: Bool = true) {
-        base.__fw_pushViewController(viewController, popWorkflows: popWorkflows, animated: animated)
+    public func push(_ viewController: UIViewController, popWorkflows: [String]?, animated: Bool = true, completion: (() -> Void)? = nil) {
+        base.__fw_pushViewController(viewController, popWorkflows: popWorkflows, animated: animated, completion: completion)
     }
 
     /// pop方式清理最外层工作流，至少保留一个根控制器（不属于工作流则不清理）
     ///
     /// 示例：1、（2、3）、4、（5、6）、（7、8），操作后为1、（2、3）、4、（5、6）
-    public func popTopWorkflow(animated: Bool = true) {
-        base.__fw_popTopWorkflow(animated: animated)
+    public func popTopWorkflow(animated: Bool = true, completion: (() -> Void)? = nil) {
+        base.__fw_popTopWorkflow(animated: animated, completion: completion)
     }
 
     /// pop方式从外到内清理指定工作流，直到遇到不属于指定工作流的控制器停止，至少保留一个根控制器
     ///
     /// 示例：1、（2、3）、4、（5、6）、（7、8），操作后为1、（2、3）、4
-    public func popWorkflows(_ workflows: [String]?, animated: Bool = true) {
-        base.__fw_popWorkflows(workflows, animated: animated)
+    public func popWorkflows(_ workflows: [String]?, animated: Bool = true, completion: (() -> Void)? = nil) {
+        base.__fw_popWorkflows(workflows, animated: animated, completion: completion)
     }
     
 }